改进微信公众号移动端网页授权的配置过程。
