+++ src/AlphaID/AuthCenterWebApp/Services/WechatMp/WechatMpAuthOptions.cs
@@ -1,4 +1,3 @@
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.OAuth;
 
 namespace AuthCenterWebApp.Services.WechatMp;
@@ -13,32 +12,5 @@ public class WechatMpAuthOptions : OAuthOptions
         Scope.Clear();
         Scope.Add("snsapi_base");
         StateDataFormat = new ShortIdStateDataFormat();
-    }
-}
-
-public class ShortIdStateDataFormat : ISecureDataFormat<AuthenticationProperties>
-{
-    public string Protect(AuthenticationProperties data)
-    {
-        var stateId = Guid.NewGuid().ToString("N");
-        WechatMpOAuthHandler.StateStore.Save(stateId, data);
-        return stateId;
-    }
-
-    public string Protect(AuthenticationProperties data, string? purpose)
-    {
-        throw new NotImplementedException();
-    }
-
-    public AuthenticationProperties? Unprotect(string? protectedText)
-    {
-        if (protectedText == null)
-            return null;
-        return WechatMpOAuthHandler.StateStore.Get(protectedText);
-    }
-
-    public AuthenticationProperties? Unprotect(string? protectedText, string? purpose)
-    {
-        throw new NotImplementedException();
     }
 }
