+++ src/AlphaID/AuthCenterWebApp/appsettings.Development.json
@@ -2,13 +2,6 @@
   "ConnectionStrings": {
     "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=AlphaIDData-62c1d690-6157-4b72-b592-86ffc498c66f;Trusted_Connection=True;MultipleActiveResultSets=True;Application Name=EntityFramework"
   },
-  "ExternalLogins": {
-    "WeixinMP": {
-      "ClientId": "wxbfb787e9643c4d49",
-      "ClientSecret": "13057ecc18f7f0cc484379531e3f961d",
-      "Enabled": true
-    }
-  },
   "IdPConfig": {
     "IssuerUri": "https://issuer.example.com"
   },
@@ -29,5 +22,11 @@
     "AdminWebAppUrl": "https://localhost:49728",
     "AuthCenterUrl": "https://localhost:49726",
     "WebApiUrl": "https://localhost:49727"
+  },
+  "WeixinMpSettings": {
+    "Enabled": true,
+    "MpName": "test-mp-name",
+    "AppId": "wxbfb787e9643c4d49",
+    "AppSecret": "13057ecc18f7f0cc484379531e3f961d"
   }
 }
