+++ src/AlphaID/AuthCenterWebApp/appsettings.json
@@ -25,11 +25,6 @@
       "ClientSecret": "<client-secret>",
       "Enabled": false
     },
-    "WeixinMP": {
-      "ClientId": "<client-id>",
-      "ClientSecret": "<client-secret>",
-      "Enabled": false
-    },
     "WorkWeixin": {
       "ClientId": "<client-id>",
       "ClientSecret": "<client-secret>",
@@ -86,5 +81,11 @@
     "AdminWebAppUrl": "<AdminWebAppUrl>",
     "AuthCenterUrl": "<AuthCenterUrl>",
     "WebApiUrl": "<WebApiUrl>"
+  },
+  "WeixinMpSettings": {
+    "Enabled": false,
+    "MpName": "<WeixinMpName>",
+    "AppId": "<AppId>",
+    "AppSecret": "<Secret>"
   }
 }
